fixup! [13.0][IMP] sale_coupon_advanced_option: Apply coupon on first sale order only
