+++ sale_coupon_advanced/__manifest__.py
@@ -1,7 +1,7 @@
 # Copyright 2020 Camptocamp
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 {
-    "name": "Sale Coupon non Cumulative Programs",
+    "name": "Sale Coupon Advanced Programs",
     "summary": "Extend coupon program features",
     "version": "13.0.1.0.0",
     "category": "Sale",
