+++ sale_coupon_advanced/readme/CONTRIBUTORS.rst
@@ -1,1 +1,2 @@
 * Iryna Vyshnevska <i.vyshnevska@mobilunity.com>
+* Anna Janiszewska <anna.janiszewska@camptocamp.com>
