+++ sale_coupon_advanced/readme/DESCRIPTION.rst
@@ -1,3 +1,4 @@
-This module extends the functionality of sale coupon to allow to create non
-cumulative program. This program will stop further computation of promotions.
+This module extends the functionality of sale coupon to allow to create:
+- non cumulative program. This program will stop further computation of promotions.
 By default all programs are applied.
+- Promotions applicable only for first sale order of the customer.
