+++ sale_coupon_advanced/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_sale_coupon_computation
+from . import test_sale_coupon_promotion_program
