+++ sale_coupon_advanced/views/sale_coupon.program.xml
@@ -9,6 +9,7 @@
         <field name="arch" type="xml">
             <field name="rule_date_to" position="after">
                 <field name="is_cumulative" />
+                <field name="first_order_only" />
             </field>
         </field>
     </record>
@@ -18,6 +19,7 @@
         <field name="arch" type="xml">
             <field name="active" position="after">
                 <field name="is_cumulative" />
+                <field name="first_order_only" />
             </field>
         </field>
     </record>
