--- sale_coupon_advanced_option/__init__.py
@@ -1,1 +1,0 @@
-from . import models
