--- sale_coupon_advanced_option/__manifest__.py
@@ -1,12 +1,0 @@
-{
-    "name": "Sale Coupon Advanced Features",
-    "summary": """Extended features for Sale Coupon""",
-    "author": "Camptocamp, Odoo Community Association (OCA)",
-    "category": "sale",
-    "version": "13.0.1.0.0",
-    "license": "AGPL-3",
-    "depends": ["sale_coupon"],
-    "data": ["views/sale_coupon.xml"],
-    "demo": [],
-    "installable": True,
-}
