--- sale_coupon_advanced_option/i18n/fr.po
@@ -1,37 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* sale_coupon_advanced_option
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 13.0+e\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2020-08-27 11:52+0000\n"
-"PO-Revision-Date: 2020-08-27 11:52+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: sale_coupon_advanced_option
-#: model:ir.model.fields,field_description:sale_coupon_advanced_option.field_sale_coupon_program__first_order_only
-msgid "Apply only first"
-msgstr "appliquer le premier"
-
-#. module: sale_coupon_advanced_option
-#: model:ir.model.fields,help:sale_coupon_advanced_option.field_sale_coupon_program__first_order_only
-msgid "Apply only on the first order of each client matching the conditions"
-msgstr "Appliquer seulement sur la première commande de chaque client répondant aux conditions"
-
-#. module: sale_coupon_advanced_option
-#: code:addons/sale_coupon_advanced_option/models/sale_coupon_program.py:0
-#, python-format
-msgid "Coupon can be used only for the first sale order!"
-msgstr "Le coupon ne peut être utilisé que pour la première commande!"
-
-#. module: sale_coupon_advanced_option
-#: model:ir.model,name:sale_coupon_advanced_option.model_sale_coupon_program
-msgid "Sales Coupon Program"
-msgstr "Programme de coupon de vente"
