--- sale_coupon_advanced_option/models/__init__.py
@@ -1,1 +1,0 @@
-from . import sale_coupon_program
