--- sale_coupon_advanced_option/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_first_so_program_only
