--- setup/sale_coupon_advanced_option/odoo/addons/sale_coupon_advanced_option
@@ -1,1 +1,0 @@
-../../../../sale_coupon_advanced_option
